# internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/sender/email/{uid}/template:
        post:
            tags:
                - Sender
            operationId: Sender_SendEmailWithTemplate
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.SendEmailWithTemplateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.SendReply'
...
    /v1/sender/webhook/{uid}/template:
        post:
            tags:
                - Sender
            operationId: Sender_SendWebhookWithTemplate
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.SendWebhookWithTemplateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.SendReply'
    /v1/template:
        post:
            tags:
                - Template
            operationId: Template_CreateTemplate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.CreateTemplateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.CreateTemplateReply'
    /v1/template/{uid}:
        get:
            tags:
                - Template
            operationId: Template_GetTemplate
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.TemplateItem'
        put:
            tags:
                - Template
            operationId: Template_UpdateTemplate
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.UpdateTemplateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.UpdateTemplateReply'
        delete:
            tags:
                - Template
            operationId: Template_DeleteTemplate
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.DeleteTemplateReply'
    /v1/template/{uid}/status:
        put:
            tags:
                - Template
            operationId: Template_UpdateTemplateStatus
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/rabbit.api.v1.UpdateTemplateStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.UpdateTemplateStatusReply'
    /v1/templates:
        get:
            tags:
                - Template
            operationId: Template_ListTemplate
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: messageType
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.ListTemplateReply'
    /v1/templates/select:
        get:
            tags:
                - Template
            operationId: Template_SelectTemplate
            parameters:
                - name: messageType
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: lastUID
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/rabbit.api.v1.SelectTemplateReply'
...
components:
    schemas:
...
        rabbit.api.v1.CreateTemplateReply:
            type: object
            properties: {}
        rabbit.api.v1.CreateTemplateRequest:
            type: object
            properties:
                name:
                    type: string
                messageType:
                    type: integer
                    format: enum
                jsonData:
                    type: string
                    description: |-
                        邮件模板数据结构:
                         {
                         	"subject": "string",
                         	"body": "string",
                         	"content_type": "string",
                         	"headers": {
                         		"key": ["value1", "value2"]
                         	}
                         }
                         SMS模板数据结构:
                         {
                         	"content": "string",
                         	"params": {
                         		"key": "value"
                         	}
                         }
                         Webhook模板数据结构:
                         {
                         }
...
        rabbit.api.v1.DeleteTemplateReply:
            type: object
            properties: {}
...
        rabbit.api.v1.ListTemplateReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/rabbit.api.v1.TemplateItem'
                total:
                    type: string
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
...
        rabbit.api.v1.SelectTemplateReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/rabbit.api.v1.TemplateItemSelect'
                total:
                    type: string
                lastUID:
                    type: string
                hasMore:
                    type: boolean
...
        rabbit.api.v1.SendEmailWithTemplateRequest:
            type: object
            properties:
                uid:
                    type: string
                templateUID:
                    type: string
                jsonData:
                    type: string
                to:
                    type: array
                    items:
                        type: string
                cc:
                    type: array
                    items:
                        type: string
...
        rabbit.api.v1.SendReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
...
        rabbit.api.v1.SendWebhookWithTemplateRequest:
            type: object
            properties:
                uid:
                    type: string
                templateUID:
                    type: string
                jsonData:
                    type: string
        rabbit.api.v1.TemplateItem:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                messageType:
                    type: integer
                    format: enum
                jsonData:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
                status:
                    type: integer
                    format: enum
        rabbit.api.v1.TemplateItemSelect:
            type: object
            properties:
                value:
                    type: string
                label:
                    type: string
                disabled:
                    type: boolean
                tooltip:
                    type: string
...
        rabbit.api.v1.UpdateTemplateReply:
            type: object
            properties: {}
        rabbit.api.v1.UpdateTemplateRequest:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                messageType:
                    type: integer
                    format: enum
                jsonData:
                    type: string
                    description: |-
                        邮件模板数据结构:
                         {
                         	"subject": "string",
                         	"body": "string",
                         	"content_type": "string",
                         	"headers": {
                         		"key": ["value1", "value2"]
                         	}
                         }
                         SMS模板数据结构:
                         {
                         	"content": "string",
                         	"params": {
                         		"key": "value"
                         	}
                         }
                         Webhook模板数据结构:
                         {
                         }
        rabbit.api.v1.UpdateTemplateStatusReply:
            type: object
            properties: {}
        rabbit.api.v1.UpdateTemplateStatusRequest:
            type: object
            properties:
                uid:
                    type: string
                status:
                    type: integer
                    format: enum
...