# pkg/api/v1/template_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v5.29.3
// source: api/v1/template.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationTemplateCreateTemplate = "/rabbit.api.v1.Template/CreateTemplate"
const OperationTemplateDeleteTemplate = "/rabbit.api.v1.Template/DeleteTemplate"
const OperationTemplateGetTemplate = "/rabbit.api.v1.Template/GetTemplate"
const OperationTemplateListTemplate = "/rabbit.api.v1.Template/ListTemplate"
const OperationTemplateSelectTemplate = "/rabbit.api.v1.Template/SelectTemplate"
const OperationTemplateUpdateTemplate = "/rabbit.api.v1.Template/UpdateTemplate"
const OperationTemplateUpdateTemplateStatus = "/rabbit.api.v1.Template/UpdateTemplateStatus"

type TemplateHTTPServer interface {
	CreateTemplate(context.Context, *CreateTemplateRequest) (*CreateTemplateReply, error)
	DeleteTemplate(context.Context, *DeleteTemplateRequest) (*DeleteTemplateReply, error)
	GetTemplate(context.Context, *GetTemplateRequest) (*TemplateItem, error)
	ListTemplate(context.Context, *ListTemplateRequest) (*ListTemplateReply, error)
	SelectTemplate(context.Context, *SelectTemplateRequest) (*SelectTemplateReply, error)
	UpdateTemplate(context.Context, *UpdateTemplateRequest) (*UpdateTemplateReply, error)
	UpdateTemplateStatus(context.Context, *UpdateTemplateStatusRequest) (*UpdateTemplateStatusReply, error)
}

func RegisterTemplateHTTPServer(s *http.Server, srv TemplateHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/template", _Template_CreateTemplate0_HTTP_Handler(srv))
	r.PUT("/v1/template/{uid}", _Template_UpdateTemplate0_HTTP_Handler(srv))
	r.PUT("/v1/template/{uid}/status", _Template_UpdateTemplateStatus0_HTTP_Handler(srv))
	r.DELETE("/v1/template/{uid}", _Template_DeleteTemplate0_HTTP_Handler(srv))
	r.GET("/v1/template/{uid}", _Template_GetTemplate0_HTTP_Handler(srv))
	r.GET("/v1/templates", _Template_ListTemplate0_HTTP_Handler(srv))
	r.GET("/v1/templates/select", _Template_SelectTemplate0_HTTP_Handler(srv))
}

func _Template_CreateTemplate0_HTTP_Handler(srv TemplateHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateTemplateRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTemplateCreateTemplate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateTemplate(ctx, req.(*CreateTemplateRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateTemplateReply)
		return ctx.Result(200, reply)
	}
}

func _Template_UpdateTemplate0_HTTP_Handler(srv TemplateHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateTemplateRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTemplateUpdateTemplate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateTemplate(ctx, req.(*UpdateTemplateRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateTemplateReply)
		return ctx.Result(200, reply)
	}
}

func _Template_UpdateTemplateStatus0_HTTP_Handler(srv TemplateHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateTemplateStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTemplateUpdateTemplateStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateTemplateStatus(ctx, req.(*UpdateTemplateStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateTemplateStatusReply)
		return ctx.Result(200, reply)
	}
}

func _Template_DeleteTemplate0_HTTP_Handler(srv TemplateHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteTemplateRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTemplateDeleteTemplate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteTemplate(ctx, req.(*DeleteTemplateRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteTemplateReply)
		return ctx.Result(200, reply)
	}
}

func _Template_GetTemplate0_HTTP_Handler(srv TemplateHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetTemplateRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTemplateGetTemplate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetTemplate(ctx, req.(*GetTemplateRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*TemplateItem)
		return ctx.Result(200, reply)
	}
}

func _Template_ListTemplate0_HTTP_Handler(srv TemplateHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListTemplateRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTemplateListTemplate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListTemplate(ctx, req.(*ListTemplateRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListTemplateReply)
		return ctx.Result(200, reply)
	}
}

func _Template_SelectTemplate0_HTTP_Handler(srv TemplateHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SelectTemplateRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTemplateSelectTemplate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SelectTemplate(ctx, req.(*SelectTemplateRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SelectTemplateReply)
		return ctx.Result(200, reply)
	}
}

type TemplateHTTPClient interface {
	CreateTemplate(ctx context.Context, req *CreateTemplateRequest, opts ...http.CallOption) (rsp *CreateTemplateReply, err error)
	DeleteTemplate(ctx context.Context, req *DeleteTemplateRequest, opts ...http.CallOption) (rsp *DeleteTemplateReply, err error)
	GetTemplate(ctx context.Context, req *GetTemplateRequest, opts ...http.CallOption) (rsp *TemplateItem, err error)
	ListTemplate(ctx context.Context, req *ListTemplateRequest, opts ...http.CallOption) (rsp *ListTemplateReply, err error)
	SelectTemplate(ctx context.Context, req *SelectTemplateRequest, opts ...http.CallOption) (rsp *SelectTemplateReply, err error)
	UpdateTemplate(ctx context.Context, req *UpdateTemplateRequest, opts ...http.CallOption) (rsp *UpdateTemplateReply, err error)
	UpdateTemplateStatus(ctx context.Context, req *UpdateTemplateStatusRequest, opts ...http.CallOption) (rsp *UpdateTemplateStatusReply, err error)
}

type TemplateHTTPClientImpl struct {
	cc *http.Client
}

func NewTemplateHTTPClient(client *http.Client) TemplateHTTPClient {
	return &TemplateHTTPClientImpl{client}
}

func (c *TemplateHTTPClientImpl) CreateTemplate(ctx context.Context, in *CreateTemplateRequest, opts ...http.CallOption) (*CreateTemplateReply, error) {
	var out CreateTemplateReply
	pattern := "/v1/template"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationTemplateCreateTemplate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *TemplateHTTPClientImpl) DeleteTemplate(ctx context.Context, in *DeleteTemplateRequest, opts ...http.CallOption) (*DeleteTemplateReply, error) {
	var out DeleteTemplateReply
	pattern := "/v1/template/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTemplateDeleteTemplate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *TemplateHTTPClientImpl) GetTemplate(ctx context.Context, in *GetTemplateRequest, opts ...http.CallOption) (*TemplateItem, error) {
	var out TemplateItem
	pattern := "/v1/template/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTemplateGetTemplate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *TemplateHTTPClientImpl) ListTemplate(ctx context.Context, in *ListTemplateRequest, opts ...http.CallOption) (*ListTemplateReply, error) {
	var out ListTemplateReply
	pattern := "/v1/templates"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTemplateListTemplate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *TemplateHTTPClientImpl) SelectTemplate(ctx context.Context, in *SelectTemplateRequest, opts ...http.CallOption) (*SelectTemplateReply, error) {
	var out SelectTemplateReply
	pattern := "/v1/templates/select"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTemplateSelectTemplate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *TemplateHTTPClientImpl) UpdateTemplate(ctx context.Context, in *UpdateTemplateRequest, opts ...http.CallOption) (*UpdateTemplateReply, error) {
	var out UpdateTemplateReply
	pattern := "/v1/template/{uid}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationTemplateUpdateTemplate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *TemplateHTTPClientImpl) UpdateTemplateStatus(ctx context.Context, in *UpdateTemplateStatusRequest, opts ...http.CallOption) (*UpdateTemplateStatusReply, error) {
	var out UpdateTemplateStatusReply
	pattern := "/v1/template/{uid}/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationTemplateUpdateTemplateStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}